Migrate visibility icon logic to StatusDetailedViewHolder
