+++ app/src/main/java/com/keylesspalace/tusky/adapter/StatusDetailedViewHolder.java
@@ -3,6 +3,7 @@ package com.keylesspalace.tusky.adapter;
 import android.content.ClipData;
 import android.content.ClipboardManager;
 import android.content.Context;
+import android.content.res.Resources;
 import android.support.annotation.Nullable;
 import android.text.SpannableStringBuilder;
 import android.text.Spanned;
@@ -11,8 +12,10 @@ import android.text.method.LinkMovementMethod;
 import android.text.style.URLSpan;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.ViewTreeObserver;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
+import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -38,6 +41,7 @@ class StatusDetailedViewHolder extends StatusBaseViewHolder {
     private TextView cardTitle;
     private TextView cardDescription;
     private TextView cardUrl;
+    private ImageView visibility;
 
     StatusDetailedViewHolder(View view) {
         super(view, false);
@@ -49,6 +53,7 @@ class StatusDetailedViewHolder extends StatusBaseViewHolder {
         cardTitle = view.findViewById(R.id.card_title);
         cardDescription = view.findViewById(R.id.card_description);
         cardUrl = view.findViewById(R.id.card_link);
+        visibility = itemView.findViewById(R.id.status_visibility);
     }
 
     @Override
@@ -154,6 +159,80 @@ class StatusDetailedViewHolder extends StatusBaseViewHolder {
             cardView.setVisibility(View.GONE);
         }
 
+        setStatusVisibility(status.getVisibility());
+    }
 
+    private void setStatusVisibility(Status.Visibility visibility) {
+        if (this.visibility == null) {
+            // The visibility icon isn't present in all status layouts, so we might not have
+            // anything to do here
+            return;
+        }
+
+        int visibilityIcon;
+        switch (visibility) {
+            case PUBLIC:
+                visibilityIcon = R.drawable.ic_public_24dp;
+                break;
+            case UNLISTED:
+                visibilityIcon = R.drawable.ic_lock_open_24dp;
+                break;
+            case PRIVATE:
+                visibilityIcon = R.drawable.ic_lock_open_24dp;
+                break;
+            case DIRECT:
+                visibilityIcon = R.drawable.ic_email_24dp;
+                break;
+            default:
+                visibilityIcon = 0;
+                ((RelativeLayout.LayoutParams) this.timestampInfo.getLayoutParams())
+                        .setMarginStart(0);
+        }
+
+        if (visibilityIcon == 0) {
+            return;
+        }
+
+        this.visibility.setImageDrawable(itemView.getContext().getDrawable(visibilityIcon));
+
+        try {
+            String[] visibilityNames = itemView.getContext()
+                    .getResources()
+                    .getStringArray(R.array.post_privacy_names);
+            if (visibilityNames.length > visibility.ordinal()) {
+                // If the strings haven't been translated to a given language, we'd throw an
+                // ArrayIndexOutOfBoundsException when trying to grab the value from the array
+                this.visibility.setContentDescription(visibilityNames[visibility.ordinal()]);
+            }
+        } catch (Resources.NotFoundException ignored) {
+            // Thrown by Resources.getStringArray()
+        }
+
+        if (this.timestampInfo == null) {
+            return;
+        }
+
+        final TextView statusTimestampInfo = this.timestampInfo;
+        final ImageView statusVisibility = this.visibility;
+        // Dynamically set the width/height of the icon to match the font size of the timestamp
+        this.timestampInfo.getViewTreeObserver().addOnGlobalLayoutListener(
+                new ViewTreeObserver.OnGlobalLayoutListener() {
+                    @Override
+                    public void onGlobalLayout() {
+                        int maxHeight = statusTimestampInfo.getHeight();
+                        if (maxHeight == 0) {
+                            // The view still hasn't been drawn, nothing to do yet
+                            return;
+                        }
+                        // The view has been drawn, so we have dimensions now
+                        statusVisibility.setMaxHeight(maxHeight);
+                        // Using the same value for width to preserve aspect ratio
+                        //noinspection SuspiciousNameCombination
+                        statusVisibility.setMaxWidth(maxHeight);
+                        statusTimestampInfo.getViewTreeObserver()
+                                .removeOnGlobalLayoutListener(this);
+                    }
+                }
+        );
     }
 }
